+++ solstice/src/main/java/dev/equo/solstice/SolsticeManifest.java
@@ -57,7 +57,21 @@ public class SolsticeManifest {
 		}
 
 		public Map<String, List<SolsticeManifest>> byExportedPackage() {
-			return groupBundles(SolsticeManifest::getPkgExports);
+			Map<String, List<SolsticeManifest>> manifests = groupBundles(SolsticeManifest::getPkgExports);
+			// fragments export the same packages as their parent, no need to report those conflicts
+			manifests.replaceAll(
+					(pkg, bundles) -> {
+						if (bundles.size() > 1) {
+							long numNonFragments =
+									bundles.stream().filter(SolsticeManifest::isNotFragment).count();
+							if (numNonFragments == 1) {
+								return Collections.singletonList(
+										bundles.stream().filter(SolsticeManifest::isNotFragment).findFirst().get());
+							}
+						}
+						return bundles;
+					});
+			return manifests;
 		}
 
 		public Map<String, List<SolsticeManifest>> calculateMissingBundles(Set<String> available) {
@@ -226,6 +240,10 @@ public class SolsticeManifest {
 		pkgImports.removeAll(pkgExports);
 	}
 
+	private boolean isNotFragment() {
+		return !headersOriginal.containsKey(Constants.FRAGMENT_HOST);
+	}
+
 	public String getSymbolicName() {
 		return symbolicName;
 	}
@@ -281,7 +299,10 @@ public class SolsticeManifest {
 			}
 			String simple = s.substring(0, attrDelim);
 			if (simple.indexOf('"') == -1) {
-				required.add(simple.trim());
+				var trimmed = simple.trim();
+				if (!required.contains(trimmed)) {
+					required.add(trimmed);
+				}
 			}
 		}
 		return required;