+++ plugin-gradle/CHANGELOG.md
@@ -6,6 +6,8 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format.
 ### Added
 - Added a parameter `useAtomos` to the Solstice buildscript, default value is `true`. ([#36](https://github.com/equodev/equo-ide/pull/36))
   - You can override the `useAtomos` buildscript parameter at the command line with `--dont-use-atomos`.
+- `equoIde` now launches the IDE in an independent process so that the build task can complete. ([#39](https://github.com/equodev/equo-ide/pull/39))
+  - Use the command line flag `--show-console` to launch the IDE as a child process and redirect its console output to the build console.
 
 ## [0.6.0] - 2022-12-16
 ### Added